Move legacy into handler source package.
