+++ mixer/pkg/runtime2/handler/env_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package legacy
+package handler
 
 import (
 	"sync"
